resources-library/apps/Discord/requirements.txt
discord.py 
python-dotenv 
PyGithub